Folder Restructuring
Community Developed - Workflows developed by the community or the
product vendor

IBM Proof Of Concept - Samples of workflows to interact with products
that are not necessarily production ready and are meant to show workflow
concepts and strategies to assist in Community Development.

IBM Verified - Workflows that are fully developed and verified by IBM

Readme files will be updated in the near future to add additional
details.

Signed-off-by: Nicolo Ereni <nicolo.ereni@nozominetworks.com>